added Category, create one to many relationship

+++ Data/Repositories/Repository.cs
@@ -38,7 +38,7 @@ namespace Data.Repositories
             return await _dbSet.ToListAsync();
         }
 
-        public async Task<TEntity> GetByIdAsync(int id)
+        public async Task<TEntity> GetByIdAsync(long id)
         {
             return await _dbSet.FindAsync(id);
         }
